* debugger: Improve GetEditorWordAtCaret to work correctly for non-rectangle selections (thanks Huki)
The old code makes a rectangle from the start of the selection to the end of
it. And it the selection is multiline and non-rectangular this code won't
return the proper string when the mouse pointer is outside the rectangle,
but still over the selected text. The new version uses one of Scintilla
functions to do it simpler and properly.